+++ CHANGELOG.md
@@ -20,11 +20,16 @@ Security in case of vulnerabilities.
  - New route `/projects/all/by-month/conversions/from/{fromMonth}/{fromYear}/to/{toMonth}/{toYear}`
  - New route `/projects/all/by-month/transfers/from/{fromMonth}/{fromYear}/to/{toMonth}/{toYear}`
  - New route `/projects/team/new`
+ - New route `/projects/team/handed-over`
  - Your team projects "new" list (`/projects/team/new`)
+ - Your team projects handed over list (`/projects/team/handed-over`)
  - Add new `ProjectTeam` extension method `TeamIsRegionalCaseworkServices`, to identify RCS users 
+  - Projects added by you (`/projects/yours/added-by`)
 
-### Changes
+### Changed
+ - Merged ListAllProjectsByFilter into main ListAllProjects query
 - Add an "orderBy" argument to the `ListAllProjectsByFilter` query
+- Allow `ListAllProjectsByFilter` query to handle multiple filters
  
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-04-24.175...main) for everything awaiting release
 
@@ -34,6 +39,7 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
 ### Added
  - New route `/projects/team/completed`
  - Your team projects completed list (`/projects/team/completed`)
+
 
 ### Changed
  - Filter out any local authorities with no projects in `ListAllProjectByLocalAuthorities`